+++ tests/test_visitors/test_ast/test_functions/test_complex_default_values.py
@@ -9,31 +9,17 @@ from wemake_python_styleguide.visitors.ast.functions import (
     FunctionDefinitionVisitor,
 )
 
-function_with_ellipse_default = """
-def __init__(self, inner_value: None = ...) -> None:
+function_with_defaults = """
+def function(self, with_default={0}):
     ...
 """
 
 
-function_with_defaults = """
-def __init__(
-    self,
-    tree,
-    filename='(none)',
-    builtins=None,
-    withDoctest={0},
-    tokens=(),
-):
-    pass
-"""
-
-bad_default = "'PYFLAKES_DOCTEST' in os.environ"
-good_default = 'SHOULD_USE_DOCTEST'
-
-
 @pytest.mark.parametrize('code', [
-    function_with_defaults,
-    bad_default,
+    "'PYFLAKES_DOCTEST' in os.environ",
+    'call()',
+    'index[1]',
+    'compare == 1',
 ])
 def test_wrong_function_defaults(
     assert_errors,
@@ -44,7 +30,7 @@ def test_wrong_function_defaults(
     mode,
 ):
     """Testing that wrong function defaults are forbidden."""
-    tree = parse_ast_tree(function_with_defaults.format(bad_default))
+    tree = parse_ast_tree(function_with_defaults.format(code))
 
     visitor = FunctionDefinitionVisitor(default_options, tree=tree)
     visitor.run()
@@ -53,8 +39,14 @@ def test_wrong_function_defaults(
 
 
 @pytest.mark.parametrize('code', [
-    function_with_defaults,
-    good_default,
+    "'string'",
+    "b''",
+    '1',
+    '-0',
+    'variable',
+    '(1, 2)',
+    'None',
+    '...',
 ])
 def test_correct_function_defaults(
     assert_errors,
@@ -65,23 +57,7 @@ def test_correct_function_defaults(
     mode,
 ):
     """Testing that correct function defaults passes validation."""
-    tree = parse_ast_tree(mode(function_with_defaults.format(good_default)))
-
-    visitor = FunctionDefinitionVisitor(default_options, tree=tree)
-    visitor.run()
-
-    assert_errors(visitor, [])
-
-
-def test_ellipse_correct_default(
-    assert_errors,
-    assert_error_text,
-    parse_ast_tree,
-    default_options,
-    mode,
-):
-    """Testing that ellipse is accepted as a correct default value."""
-    tree = parse_ast_tree(mode(function_with_ellipse_default))
+    tree = parse_ast_tree(mode(function_with_defaults.format(code)))
 
     visitor = FunctionDefinitionVisitor(default_options, tree=tree)
     visitor.run()